backend/requirements.txt
asgiref==3.6.0
# Django==4.2.13
GitPython==3.1.44
python-gitlab==5.6.0
PyJWT==2.10.1
PyMySQL==1.1.1
Requests==2.32.3
uvicorn==0.34.0
decorator==5.1.1
django-cors-headers==4.2.0
cryptography==42.0.5
PyYAML==6.0.1